The main resource is the default entity manager.
The main use case when getting the resource is to get the default
EntityManager. With the old code you find yourself doing this a lot:

$em = $bootstrap->getResource('Doctrine')->getEntityManager();

With this change all you need to do is:

$em = $bootstrap->getResource('Doctrine');

You can still obtain a reference to the Doctrine plugin to get other
entity managers, DBAL instances, etc... like this:

$doctrine = $bootstrap->getPluginResource('Doctrine');

Thus, no functionality is lost but the main use case has become much
simpler.